update regression predict

+++ ml/src/main/java/kr/gaion/armoredVehicle/database/model/SensorTempLife.java
@@ -22,7 +22,7 @@ public class SensorTempLife {
 //    private Double AiPredict;
 
     @Column(name="ACPOWER")
-    private Float acPower;
+    private Double acPower;
 
     @Column(name="AI_Algorithm")
     private String aiAlgorithm;
@@ -31,28 +31,28 @@ public class SensorTempLife {
     private String aiModel;
 
     @Column(name="CPUUTIL")
-    private Float cpuUtil;
+    private Double cpuUtil;
 
     @Column(name="DISKACCESSES")
-    private Float diskAccesses;
+    private Double diskAccesses;
 
     @Column(name="DISKBLOCKS")
-    private Float diskBlocks;
+    private Double diskBlocks;
 
     @Column(name="DISKUTIL")
-    private Float diskUtil;
+    private Double diskUtil;
 
     @Column(name="INSTRETIRED")
-    private Float instRetired;
+    private Double instRetired;
 
     @Column(name="LASTLEVEL")
-    private Float lastLevel;
+    private Double lastLevel;
 
     @Column(name="MEMORYBUS")
-    private Float memoryBus;
+    private Double memoryBus;
 
     @Column(name="CORECYCLE")
-    private Float coreCycle;
+    private Double coreCycle;
 
     @Column(name="TIME")
     private String time;
